Complete payment, withdrawal and transaction APIs

+++ aitasker/src/main/java/com/aitasker/job/entity/JobPost.java
@@ -6,6 +6,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -20,7 +21,7 @@ public class JobPost extends BaseEntity {
     private String title;
     @Column(columnDefinition = "TEXT")
     private String description;
-    private Double budget;
+    private BigDecimal budget;
     private LocalDate deadline;
     private String requiredSkills;
     @Enumerated(EnumType.STRING)
